rename mn_setup.h->mn_internal.h
