+++ src/mn_internal.h
